% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{roc}
\alias{proproc}
\title{Deprecated Functions}
\usage{
roc(truth, rating, ...)

proproc(truth, rating, ...)
}
\arguments{
\item{truth, rating, ...}{arguments passed to non-deprecated equivalent.}
}
\description{
Functions that have been deprecated and will be removed in the future.
}
\details{
Use \code{\link[=roc_curves]{roc_curves()}} instead of \code{roc()}.

Use \code{\link[=roc_curves]{roc_curves()}} instead of \code{proproc()}.
}
