% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kundel}
\alias{Kundel}
\title{Multi-reader multi-case dataset}
\format{
A data frame with 760 rows and 5 variables:
\describe{
  \item{reader}{reader identifier}
  \item{treatment}{numeric identifier for treatment modality}
  \item{treatment_name}{name of treatment modality}
  \item{case}{patient identifier}
  \item{truth}{true case status (1 = diseased, 0 = non-diseased)}
  \item{rating}{ordinal reader ratings of case status}
}
}
\usage{
Kundel
}
\description{
Data from a multi-reader multi-case study conducted by Kundel et al. (1997)
to compare the clinical diagnostic accuracy of hard-copy readings of
screen-film bedside chest radiographs and both hard- and soft-copy readings
of computed radiographs from separate groups of patients in a medical
intensive care unit.
}
\references{
Kundel HL, Gefter W, Aronchick J, Miller W Jr, Hatabu H, Whitfill CH, Miller
W Sr (1997). Accuracy of bedside chest hard-copy screen-film versus hard- and
soft-copy computed radiographs in a medical intensive care unit: receiver
operating characteristic analysis. Radiology, 205(3): 859-63.
}
\keyword{datasets}
