% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmc.R
\name{mrmc}
\alias{mrmc}
\title{Multi-Reader Multi-Case ROC Analysis}
\usage{
mrmc(formula, cases, data, method = DeLong)
}
\arguments{
\item{formula}{formula defining the response metric and variables on the left
hand side and the test and reader variables on the right hand side.
The response should be expressed in terms of a package-supplied performance
\code{\link[mrROC:metrics]{metric}}.  Test and reader variables should be specified in that
order and separated by the \code{+} operator.}

\item{cases}{variable containing the case identifiers.}

\item{data}{data frame containing the \code{formula} and \code{cases}
variables.}

\item{method}{function, function call, or character string naming the
\code{\link[mrROC:cov_methods]{method}} to use in calculating performance
metric covariances.}
}
\description{
Estimation and comparison of ROC performance metrics for multi-reader
multi-case studies.
}
\examples{
(est <- mrmc(roc_auc(truth, rating) ~ treatment + reader, cases = case,
             data = VanDyke))
summary(est)
plot(est)

}
\seealso{
\code{\link{metrics}}, \code{\link{cov_methods}},
\code{\link{summary}}, \code{\link{plot}}
}
