% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{OR_to_RM}
\alias{OR_to_RM}
\alias{OR_to_RM.default}
\alias{OR_to_RM.data.frame}
\alias{OR_to_RM.list}
\title{Convert Obuchowski-Rockette to Roe & Metz Parameters}
\usage{
OR_to_RM(...)

\method{OR_to_RM}{default}(
  AUC1,
  AUC2,
  var_R,
  var_TR,
  corr1,
  corr2,
  corr3,
  error_var,
  n0,
  n1,
  precision = 1e-04,
  b_option = 1,
  mean_to_sig = 4.5,
  b_input = 1,
  ...
)

\method{OR_to_RM}{data.frame}(params, ...)

\method{OR_to_RM}{list}(params, ...)
}
\arguments{
\item{...}{arguments passed to the default method.}

\item{AUC1, AUC2}{test 1 and 2 AUCs.}

\item{var_R, var_TR}{OR reader and test-by-reader variances.}

\item{corr1, corr2, corr3}{OR fixed reader correlations.}

\item{error_var}{OR fixed reader error variance.}

\item{n0, n1}{number of nondiseased and diseased cases.}

\item{precision}{precision for estimation of RM using numerical methods.}

\item{b_option}{method of estimating RM binormal b parameter (= 1, 2, or 3).}

\item{mean_to_sig}{mean-to-sigma ratio (Hillis & Berbaum, 2011), required
only if \code{b_option = 2}.}

\item{b_input}{binormal b value, required only if \code{b_option = 3}.}

\item{params}{dataframe of OR parameter values in the columns or a list of
OR parameter values.}
}
\value{
The RM model parameters and relevant inputs are returned as a data frame with
the following elements.
\describe{
  \item{n0}{}
  \item{n1}{}
  \item{b}{}
  \item{mu1}{}
  \item{mu2}{}
  \item{var_R}{}
  \item{var_TR}{}
  \item{var_C}{}
  \item{var_TC}{test-by-case variance.}
  \item{var_RC}{reader-by-case variance.}
  \item{var_error}{}
  \item{b_option}{}
}
}
\description{
Determines Roe & Metz (RM) simulation model parameters for simulating
multireader multicase likelihood-of-disease rating data based on real-data or
conjectured Obuchowski-Rockette (OR) parameter estimates that describe the
distribution of the empirical AUC reader performance measure. The algorithm
is for the constrained unequal-variance RM model (Hillis, 2012), which
generalizes the orginal RM model (Roe and Metz, 1997) by allowing the
diseased and nondiseased decision-variable distributions to have unequal
variances for each reader.  Parameter values for the the original RM model
can be computed by setting \code{b_option = 3} and \code{b_input = 1}.
}
\details{
Hillis (2018) derived formulas expressing OR parameters that describe the
distribution of epirical AUC outcomes computed from RM simulated data as
functions of the RM parameters.  This algorithm provides the reverse
transformation that determines the corresponding RM parameters.  The
algorithm is described in Hillis (2020).


\code{precision} indicates the numerical accuracy of the algorithm.
Typically, values less than 0.0001 will result in little difference.  For
final results we recommend using 0.00001, with smaller values (e.g., 0.001,
0.0001) used for preliminary results to speed up the algorithm.

\code{b_option} indicates the method for estimating the RM binormal \emph{b}
parameter. \code{b_option = 1} results in RM simulated data for which
expected OR estimates are equal to those specified in \code{params}.
\code{b_option = 2} results in RM parameters such that the median
mean-to-sigma ratio across readers for the RM binormal distributions
corresponding to the lowest OR AUC is equal to \code{mean-to_sig}.
\code{b_option = 3} allows the user to input the desired value of \emph{b}
(e.g., set \code{b_input = 1} to obtain original RM parameter estimates.
For \code{b_option = 2,3}, the simulated data empirical AUC estimate
distribution is specified by the parameter values specified in
\code{OR_params}, \emph{except} for error_var_OR. Thus for b_options 2 & 3,
\code{error_var_OR} can be equal to \code{NA} or excluded from \code{params}.

\code{mean_to_sig} is the inputted mean-to-sigma ratio needed for
\code{b_option = 2}.

\code{b_input} is the inputted \emph{b} value needed for \code{b_option = 3}.

The OR model includes the following constraints:  \code{corr1_OR >= corr3_OR}
and \code{corr2_OR >= corr3_OR}. These constraints are  included in the
algorithm by setting \code{corr2_OR = corr3_OR} if \code{corr2_OR < corr3_OR}
(and similarly for \code{corr1_OR} if \code{corr1_OR < corr3_OR}).
}
\examples{
## Example 1a: Computing RM parameters from OR parameters directly
RM <- OR_to_RM(n0 = 69, n1 = 45, AUC1 = 0.897, AUC2 = 0.941,
               corr1 = 0.433, corr2 = 0.430, corr3 = 0.299,
               var_R = 0.00154, var_TR = 0.000208, error_var = 0.000788,
               precision = 0.0001)
RM

## Example 1b: Computing RM parameters from a list of OR parameters
vandyke_OR <- list(n0 = 69, n1 = 45, AUC1 = 0.897, AUC2 = 0.941,
                   corr1 = 0.433, corr2 = 0.430, corr3 = 0.299,
                   var_R = 0.00154, var_TR = 0.000208, error_var = 0.000788)
vandyke_RM <- OR_to_RM(vandyke_OR, precision = 0.0001)
vandyke_RM

## Example 1c: Computing RM parameters from a data frame of OR parameters
three_studies_OR <- data.frame(
  rbind(
    vandyke = c(69, 45, 0.89793704, 0.94083736, 0.432, 0.429, 0.298, 0.00154,
                0.0002, 0.00080229),
    franken = c(67, 33, .8477498869, 0.8368950701, 0.521430051, 0.319691199,
                0.3386375697, 0.0000433385, 0.0, 0.0014967254),
    kundel = c(66, 29, 0.8038793103, 0.8413662487, 0.507695244, 0.3843523762,
               0.4035662578, 0.0007340122, 0, 0.002148844)
  )
)
colnames(three_studies_OR) <- c("n0", "n1", "AUC1", "AUC2", "corr1", "corr2",
                                "corr3", "var_R", "var_TR", "error_var")
three_studies_RM <- OR_to_RM(three_studies_OR, precision = 0.0001)
three_studies_RM

}
\references{
Hillis, Stephen. 2020. "Determining Roe and Metz model parameters for
simulating multireader multicase confidence-of-disease rating data based on
read-data or conjectured Obuchowski-Rockette parameter estimates." Vol.
11316, SPIE Medical Imaging: SPIE. doi.org/10.1117/12.2550541

Hillis, Stephen L. 2012. "Simulation of Unequal-Variance Binormal Multireader
ROC Decision Data: An Extension of the Roe and Metz Simulation Model."
\emph{Academic Radiology} no. 19 (12):1518-1528.
doi: 10.1016/j.acra.2012.09.011.

Hillis, Stephen L. 2018. "Relationship between Roe and Metz simulation model
for multireader diagnostic data and Obuchowski-Rockette model parameters."
\emph{Statistics in Medicine} no. 37 (13):2067-2093. doi: 10.1002/sim.7616.

Roe, Cheryl A., and Charles E. Metz. 1997. "Dorfman-Berbaum-Metz method for
statistical analysis of multireader, multimodality receiver operating
characteristic data: Validation with computer simulation." \emph{Academic
Radiology} no. 4 (4):298-303. doi: doi: 10.1016/S1076-6332(97)80032-3.
}
\author{
Stephen L. Hillis, Departments of Radiology and Biostatistics,
University of Iowa, steve-hillis@uiowa.edu
}
