% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.roc_curve}
\alias{print.roc_curve}
\alias{print.roc_curves}
\title{Print ROC Objects}
\usage{
\method{print}{roc_curve}(x, n = 11, ...)

\method{print}{roc_curves}(x, n_curves = 5, n = 11, ...)
}
\arguments{
\item{x}{object to print.}

\item{n}{number of ROC curve points to print.}

\item{...}{arguments passed to other methods.}

\item{n_curves}{number of ROC curves to print.}
}
\description{
Print ROC objects from the \pkg{MRMCaov} package.
}
\examples{
curves <- with(VanDyke,
  roc_curves(truth, rating, groups = list(Test = treatment, Reader = reader))
)
print(curves)

}
