% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\alias{proproc}
\title{ROC Performance Metrics}
\usage{
roc(truth, rating, ...)

proproc(truth, rating, ...)
}
\arguments{
\item{truth}{vector of true binary statuses.}

\item{rating}{vector of numeric ratings.}

\item{...}{interaction factors within which to calculate metrics.}
}
\description{
Calculation of TPR and FPR pairs for all values of a numeric rating of a
true binary response.
}
\examples{
with(VanDyke, roc(truth, rating, treatment, reader))

}
\seealso{
\code{\link{plot}}
}
