% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srmc.R
\name{srmc}
\alias{srmc}
\title{Single-Reader Multi-Case ROC Analysis}
\usage{
srmc(response, case, data, method = jackknife)
}
\arguments{
\item{response}{response metric expressed in terms of a package-supplied
performance \code{\link[=metrics]{metric}}.}

\item{case}{optional variable of case identifiers.}

\item{data}{data frame containing the \code{response}, \code{test},
\code{reader}, and \code{case} variables.}

\item{method}{function, function call, or character string naming the
\code{\link[=cov_methods]{method}} to use in calculating performance
metric covariances.}
}
\description{
Estimation of ROC performance metrics for a single reader of multiple cases.
}
\examples{
VanDyke1 <- subset(VanDyke, reader == 1)
est <- srmc(empirical_auc(truth, rating), data = VanDyke1)
summary(est)

}
\seealso{
\code{\link{metrics}}, \code{\link{cov_methods}},
\code{\link{parameters}}, \code{\link{plot}}, \code{\link{roc_curves}},
\code{\link{summary}}
}
