% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmc.R
\name{summary}
\alias{summary}
\alias{summary.mrmc}
\title{Summary Estimates and Statistical Tests}
\usage{
\method{summary}{mrmc}(object, design = 1, conf.level = 0.95, ...)
}
\arguments{
\item{object}{object to summarize.}

\item{design}{one of the following study designs: 1 = factorial, 2 = cases
nested within readers, or 3 = cases nested within tests.}

\item{conf.level}{confidence level for confidence intervals.}

\item{...}{additional arguments affecting the summary.}
}
\description{
Summary Estimates and Statistical Tests
}
\seealso{
\code{\link{mrmc}}
}
