% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeLong.R, R/jackknife.R, R/unbiased.R
\name{cov_methods}
\alias{cov_methods}
\alias{DeLong}
\alias{jackknife}
\alias{unbiased}
\title{Performance Metric Covariance Estimation}
\usage{
DeLong()

jackknife()

unbiased(abar = FALSE)
}
\arguments{
\item{abar}{logical indicating whether to compute mean covariance components.}
}
\description{
Performance Metric Covariance Estimation
}
\seealso{
\code{\link{mrmc}}
}
