% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{binormroc_auc}
\alias{proproc_auc}
\alias{roc_auc}
\title{Performance Metrics}
\usage{
binormroc_auc(truth, rating)

proproc_auc(truth, rating)

roc_auc(truth, rating)
}
\arguments{
\item{truth}{vector of true binary statuses.}

\item{rating}{vector of numeric ratings.}
}
\description{
Performance Metrics
}
\details{
Performance metrics measure the degree to which higher case ratings are
associated with positive case statuses, where positive status is taken to be
the highest level of \code{truth}.  Available metrics include the following.
\describe{
\item{binormroc_auc:}{Area under the binormal ROC curve.}
\item{proproc_auc:}{Area under a proper ROC curve.}
\item{roc_auc:}{Area under the empirical, or trapezoidal, ROC curve.}
}
}
