% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{binary_sens}
\alias{binary_spec}
\alias{binormal_auc}
\alias{binormal_sens}
\alias{binormal_spec}
\alias{empirical_auc}
\alias{empirical_sens}
\alias{empirical_spec}
\alias{proproc_auc}
\alias{proproc_sens}
\alias{proproc_spec}
\alias{trapezoidal_auc}
\alias{trapezoidal_sens}
\alias{trapezoidal_spec}
\title{Performance Metrics}
\usage{
binary_sens(truth, rating)

binary_spec(truth, rating)

binormal_auc(truth, rating, partial = FALSE, min = 0, max = 1)

binormal_sens(truth, rating, spec)

binormal_spec(truth, rating, sens)

empirical_auc(truth, rating, partial = FALSE, min = 0, max = 1)

empirical_sens(truth, rating, spec)

empirical_spec(truth, rating, sens)

proproc_auc(truth, rating, partial = FALSE, min = 0, max = 1)

proproc_sens(truth, rating, spec)

proproc_spec(truth, rating, sens)

trapezoidal_auc(truth, rating, partial = FALSE, min = 0, max = 1)

trapezoidal_sens(truth, rating, spec)

trapezoidal_spec(truth, rating, sens)
}
\arguments{
\item{truth}{vector of true binary statuses.}

\item{rating}{vector of binary ratings for the binary metrics and ranges of
numeric ratings for the others.}

\item{partial}{character string \code{"sensitivity"} or \code{"specificity"}
for calculation of partial AUC, or \code{FALSE} for full AUC.  Partial
matching of the character strings is allowed.}

\item{min, max}{minimum and maximum sensitivity or specificity values over
which to calculate partial AUC.}

\item{sens, spec}{numeric sensitivity/specificity at which to calculate
specificity/sensitivity.}
}
\description{
Estimated performance metrics from ROC curves.
}
\details{
Performance metrics measure the degree to which higher case ratings are
associated with positive case statuses, where positive status is taken to be
the highest level of \code{truth}.  Available metrics include the following.
\describe{
\item{empirical_auc, trapezoidal_auc:}{Area under the empirical, or
trapezoidal, ROC curve.}
\item{proproc_auc:}{Area under a proper ROC curve.}
}
}
