% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{binormroc_auc}
\alias{proproc_auc}
\alias{roc_auc}
\title{Performance Metrics}
\usage{
binormroc_auc(truth, rating)

proproc_auc(truth, rating)

roc_auc(truth, rating)
}
\arguments{
\item{truth}{vector of true binary responses.}

\item{rating}{vector of ratings.}
}
\description{
Performance Metrics
}
