% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmc.R
\name{mrmc}
\alias{mrmc}
\title{Multi-Reader Multi-Case ROC Analysis}
\usage{
mrmc(response, test, reader, case, data, method = jackknife, design = NULL)
}
\arguments{
\item{response}{response metric expressed in terms of a package-supplied
performance \code{\link[=metrics]{metric}}.}

\item{test}{variable of test identifiers.}

\item{reader}{variable of reader identifiers.}

\item{case}{variable of case identifiers.}

\item{data}{data frame containing the \code{response}, \code{test},
\code{reader}, and \code{case} variables.}

\item{method}{function, function call, or character string naming the
\code{\link[=cov_methods]{method}} to use in calculating performance
metric covariances.}

\item{design}{one of the following study designs: 1 = factorial, 2 = cases
nested within readers, 3 = cases nested within tests, or \code{NULL} to
automatically set the design based on variable codings in data.}
}
\description{
Estimation and comparison of ROC performance metrics for multi-reader
multi-case studies.
}
\details{
Readers and cases are treated as random factors by default.  Either one may
be designated as fixed in calls to \code{mrmc} with the syntax
\code{fixed(<variable name>)}, where \code{<variable name>} is the name of
the reader or case variable.
}
\examples{
## Random readers and cases
(est <- mrmc(empirical_auc(truth, rating), treatment, reader, case,
             data = VanDyke))
summary(est)
plot(est)

## Fixed readers and fixed cases
est <- mrmc(empirical_auc(truth, rating), treatment, fixed(reader), case,
            data = VanDyke)
summary(est)

}
\seealso{
\code{\link{metrics}}, \code{\link{cov_methods}},
\code{\link{parameters}}, \code{\link{plot}}, \code{\link{roc_curves}},
\code{\link{summary}}
}
