% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot.roc}
\alias{plot.mrmc}
\title{Performance Plots}
\usage{
\method{plot}{roc}(x, ...)

\method{plot}{mrmc}(x, ...)
}
\arguments{
\item{x}{object to plot.}

\item{...}{arguments passed to other methods.}
}
\description{
Performance Plots
}
\examples{
perf <- with(VanDyke, roc(truth, rating, treatment, reader))
plot(perf)

}
\seealso{
\code{\link{roc}}
}
