% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot.roc_curve}
\alias{plot.roc_curves}
\alias{plot.empirical_curve}
\alias{plot.empirical_curves}
\alias{plot.roc_points}
\alias{plot.mrmc}
\title{Performance Plots}
\usage{
\method{plot}{roc_curve}(x, n = 100, ...)

\method{plot}{roc_curves}(x, n = 100, ...)

\method{plot}{empirical_curve}(x, ...)

\method{plot}{empirical_curves}(x, ...)

\method{plot}{roc_points}(x, ...)

\method{plot}{mrmc}(x, n = 100, ...)
}
\arguments{
\item{x}{object to plot.}

\item{n}{number of equally spaced false-positive rate points at which to
calculate true-positive rates and interpolate through for display of the
curve.}

\item{...}{arguments passed to other methods.}
}
\description{
Performance Plots
}
\examples{
curves <- with(VanDyke,
  roc_curves(truth, rating, groups = list(Test = treatment, Reader = reader))
)
plot(curves)

}
\seealso{
\code{\link{roc_curves}}
}
