% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{roc}
\alias{roc}
\title{ROC Performance Metrics}
\usage{
roc(observed, predicted, ...)
}
\arguments{
\item{observed}{vector of observed binary responses.}

\item{predicted}{vector of predcited responses.}

\item{...}{interaction factors within which to calculate metrics.}
}
\description{
Calculation of TPR and FPR pairs for all values of a numeric predictor
of an observed binary response.
}
\examples{
with(VanDyke, roc(truth, rating, treatment, reader))

}
\seealso{
\code{\link{plot}}
}
