% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc_curves}
\alias{roc_curves}
\alias{points.empirical_curves}
\alias{points.proproc_curves}
\alias{mean.roc_curves}
\title{ROC Performance Curves}
\usage{
roc_curves(truth, rating, groups = list(), method = "empirical")

\method{points}{empirical_curves}(x, values = NULL, metric = c("specificity", "sensitivity"), ...)

\method{points}{proproc_curves}(
  x,
  values = seq(0, 1, length = 100),
  metric = c("specificity", "sensitivity"),
  ...
)

\method{mean}{roc_curves}(x, ...)
}
\arguments{
\item{truth}{vector of true binary statuses.}

\item{rating}{vector of numeric ratings.}

\item{groups}{list or data frame of grouping variables of the same lengths as
\code{truth} and \code{rating}.}

\item{method}{character string \code{"empirical"}, \code{"trapezoidal"}, or
\code{"proproc"} indicating the curve type.}

\item{x}{object returned by \code{roc_curves} for which to compute points on
or to average over the curves.}

\item{values}{numeric vector of values at which to compute the points.  If
\code{NULL} then the intersection of all empirical points is used.}

\item{metric}{reader performance metric to which the \code{values}
correspond.}

\item{...}{arguments passed from the \code{mean()} method to \code{points()}.}
}
\description{
Calculation of TPR and FPR pairs for values of a numeric rating of a
true binary response.
}
\examples{
curves <- with(VanDyke,
  roc_curves(truth, rating, groups = list(Test = treatment, Reader = reader))
)
points(curves)
mean(curves)

}
\seealso{
\code{\link{plot}}
}
