% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.mrmc}
\title{Summary Estimates and Statistical Tests}
\usage{
\method{summary}{mrmc}(object, conf.level = 0.95, ...)
}
\arguments{
\item{object}{object to summarize.}

\item{conf.level}{confidence level for confidence intervals.}

\item{...}{additional arguments affecting the summary.}
}
\description{
Summary Estimates and Statistical Tests
}
\seealso{
\code{\link{mrmc}}
}
