% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VanDyke}
\alias{VanDyke}
\title{Multi-reader multi-case dataset}
\format{A data frame with 1140 rows and 6 variables:
\describe{
  \item{reader}{reader identifier}
  \item{treatment}{treatment identifier}
  \item{case}{case identifier}
  \item{truth}{true case status (1 = positive, 0 = negative)}
  \item{rating}{ordinal reader ratings of case status}
}}
\usage{
VanDyke
}
\description{
Multi-reader multi-case dataset
}
\keyword{datasets}
